Debounce based on message
I scaled an app twice within a minute, and the second event got
debounced even though it was real. Now we deboune based on the last
message we sent, which will be more accurate. Having the message stored
on the event is handy for a DB view too.

Since the Event now contains its own message, SlackNotifier no longer
needs to know what the Notifications are.